+++ codemods/custom-property-replacer/README.md
@@ -2,11 +2,7 @@
 
 The purpose of this PostCSS plugin is to replace deprecated custom properties with the new custom properties. To use this plugin, follow the following instructions.
 
-First, start by installing `postcss-cli`. You will use this to run the css transformation.
-
-`yarn add -D postcss-cli`
-
-Next, create a file called `postcss.config.js` in the root of your project.
+First create a file called `postcss.config.js` in the root of your project.
 
 ```js
 module.exports = {
@@ -14,7 +10,7 @@ module.exports = {
     require('@commercetools-frontend/ui-kit/codemods/custom-property-replacer')(
       {
         file: require.resolve(
-          '@commercetools-frontend/ui-kit/codemods/custom-property-replacer/variable-mapping.json'
+          '@commercetools-frontend/ui-kit/codemods/custom-property-replacer/v10/variable-mapping.json'
         ),
       }
     ),
@@ -26,6 +22,6 @@ Note: if you are using SCSS instead of CSS, you may need to change the parser. P
 
 Next, use the command line to run the code transformation.
 
-`yarn postcss postcss-codemods/**/*.css -r`
+`npx postcss-cli **/*.css -r`
 
-This should replace your usage of deprecated custom properties with the new custom properties.
+This will replace your usage of deprecated custom properties with the new custom properties.
