+++ codemods/custom-property-replacer/v10/variable-mapping.json
